requirements.txt
ncatbot~=3.8.10.post3
requests~=2.32.4
beautifulsoup4~=4.13.4
pyjwt~=2.10.1
aiohttp~=3.12.15
geopy~=2.4.1
toml~=0.10.2